move txLink & addressLink into common file
